+++ src/main/java/roomescape/user/service/TokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape.user;
+package roomescape.user.service;
 
 public interface TokenProvider {
